V0.40.0 - Talk To NPC Quest * Added a quest to handle talking to NPCs * Quest can hold multiple NPCs that each can have an "infinite" number of dialog strings to give to the player. + Changed methods of adding XP and Money from setMoney(mc.getMoney() + money) to addMoney(money) and same for XP - Removed the saving of Location quests. There is no need to save that since the location is a single entity that is stored in the quest on quest creation.
Also, will start working on a better way to handle the saving/loading/creation of quests. Currently, it's pretty tedious to go through and create a quest. Requires quite a bit of code. But I want an API-type framework to be able to register quests, save the quest information, and load the quest information from the player. Instead of having to deal with manually typing that in each time I want it to be saved/loaded/created.