V0.53.0 - Quest API and Visuals * Added new method for quicker and easier use of creating the ItemStack for the quests + Added nice visual updates for in game when looking at quests + Adding some works for CustomWeapons
(BUG): If you leave default requiresTurnIn = true, and got a quest from a quest. (As in the reward for automatic quest giving). The player will not be able to complete the quest because of how the quest line turned. Will need to find a way to make sure the quest can be turned in to the NPC even if it was given from a quest.