V0.54.0 - ArmorStandMob removal
The CustomArmorStandMob package has been removed and will be worked on for better usability. Currently, there is lots of issues with it. One, must be custom coded as of right now, and had limitations on things like textures, sizes, and more. It just wasn't cutting it. I will be moving to a new method. This new method will be using BlockBench custom made objects and using them to create custom item models. Which, can make entire custom mobs and allow people to create their own mobs using the BlockBench creator as well as textures. This will then put all required items into a texture pack for the server owner to install on their server and then allow for players to see amazing custom mobs.