V0.68.0 - Major Quest Rework * getParentQuestsForQuest(String) was added to QuestManager to account for not being able to get the Quest object easily. + Quest unicode values turned into UTF-8 characters instead + NPC givableQuest list is not a String list not a Quest list. + findNpcWithQuest now takes a Quest name not a Quest object + MMOPlayer has a new rework of method addQuest(String). It will now take a quest name, add it to the current MMOClass, and then will return the quest as well to be used in other functions.
With the rework of V0.66.0, Quests were changed a bit to be able to account for strings mainly. This apparently caused a huge issue when trying to remove the quest from the players CompletedableQuest list when they completed a quest causing them to be able to infinitely collect rewards for a quest that was already completed since it wasn't being removed from there inventory. This fix accounts for that, turning almost all but the ActiveQuest list into name placeholders of the quest. As well as addressing that bug of not removing the quest from the players quest list.