V0.71.0 - Dungeon Start
A MMO-RPG is not complete without some form of Instance/Dungeon style of gameplay. The goal is to be able to allow people to create "blocks" and have an auto-generated procedually-generated dungeon from the DungeonBlocks gathered in the schematics folder. As the Dungeon gets built out, triggers will be available. Planned triggers will be: Physical (Pressure plate, Button, Door w/ Key), Kill (Once everything in the room is killed), Room Enter, Room Exit, and "Floor" complete. Allowing for a lot of configuration and customizability when making instances. Linking special files with the schematics.