V0.73.0 - Dungeon Block Pasting
* Added new command "schematic" used to test out pasting schematics into the world
* Added a pasetAll() function into Dungeon to go through each of the DungeonBlocks and paste them into the world at the location given
* Added "pasteBlock()" to read all of the DungeonSchematicBlockInformation from the DungeonBlock and grab it's material, location, and blockdata and paste it into the world
+ Fixed reading material strings, it was adding ":" into the material string when it shouldn't have
+ Fixed if the block material string doesn't have brackets for the BlockData

As this goes forward, there will be lots of reworks and bugs as this plugin develops. This will most likely be the hardest part plugin as I have to make a dungeon creator that will place random rooms together with walkways and incorporate the ability to select specific blocks in the DungeonBlock that can trigger events.