V0.85.0 - Player Loading Fix * MaxHealth added to MMOClass saving/loading * Mana added to MMOClass saving/loading
There was an issue leaving the server after you had selected your class and rejoin. This was caused with the loading of the player looking for "MaxHealth" and "Mana" String that wasn't present. This was because I had forgot that I already added this when I stopped developing a little while ago. This has now been fixed