V0.95.0 - Stats Updating
After a little bit of trouble shooting, I fixed some code that I had pulled from an old plugin long ago. This code was making it so that getModsOn.put() was ONLY being called if there was no CraftingRarity. The new code, ALL items already have a CraftingRarity. By getting rid of that if statement, stats are being updated now. But, I ran into a problem. Currently, the stats are updated AFTER you move the item. If I take an item off, it is updating the stats for the item as if it was there. If I am putting an item on, it is taking the stats of the item as if it was not there.